# Add benchmark to measure CustomBinaryDocValuesField performance