@kkmr 2023-08-27
New code Performance

```
Benchmark                                        (entrySize)  (maximumNumberOfEntries)   Mode  Cnt     Score   Error   Units
CustomBinaryDocValuesFieldBenchmark.add                    8                         8  thrpt       4826.142          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                    8                        32  thrpt       1380.164          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                    8                       128  thrpt        358.869          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                    8                       512  thrpt         90.848          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                   32                         8  thrpt       2506.266          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                   32                        32  thrpt        662.682          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                   32                       128  thrpt        168.800          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                   32                       512  thrpt         40.062          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                  128                         8  thrpt        742.890          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                  128                        32  thrpt        188.871          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                  128                       128  thrpt         47.518          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                  128                       512  thrpt         11.920          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                  512                         8  thrpt        210.733          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                  512                        32  thrpt         53.659          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                  512                       128  thrpt         13.238          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                  512                       512  thrpt          3.352          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue            8                         8  thrpt       9542.207          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue            8                        32  thrpt       9916.014          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue            8                       128  thrpt       9466.273          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue            8                       512  thrpt       9323.077          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue           32                         8  thrpt       9253.423          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue           32                        32  thrpt       9535.728          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue           32                       128  thrpt       9156.015          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue           32                       512  thrpt       9203.324          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue          128                         8  thrpt       7356.562          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue          128                        32  thrpt       7757.455          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue          128                       128  thrpt       7553.087          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue          128                       512  thrpt       7496.237          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue          512                         8  thrpt       5050.827          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue          512                        32  thrpt       5048.714          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue          512                       128  thrpt       4975.164          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue          512                       512  thrpt       5085.280          ops/ms
```

Old code performance

```
Benchmark                                        (entrySize)  (maximumNumberOfEntries)   Mode  Cnt      Score   Error   Units
CustomBinaryDocValuesFieldBenchmark.add                    8                         8  thrpt        9012.270          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                    8                        32  thrpt        2002.192          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                    8                       128  thrpt         571.450          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                    8                       512  thrpt         153.650          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                   32                         8  thrpt        3256.273          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                   32                        32  thrpt         735.980          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                   32                       128  thrpt         201.739          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                   32                       512  thrpt          52.079          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                  128                         8  thrpt         807.506          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                  128                        32  thrpt         190.050          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                  128                       128  thrpt          49.815          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                  128                       512  thrpt          12.696          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                  512                         8  thrpt         215.063          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                  512                        32  thrpt          52.353          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                  512                       128  thrpt          13.426          ops/ms
CustomBinaryDocValuesFieldBenchmark.add                  512                       512  thrpt           3.326          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue            8                         8  thrpt       13037.574          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue            8                        32  thrpt       12014.520          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue            8                       128  thrpt       13252.598          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue            8                       512  thrpt       12223.060          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue           32                         8  thrpt       13465.619          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue           32                        32  thrpt       13116.687          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue           32                       128  thrpt       13456.720          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue           32                       512  thrpt       12030.096          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue          128                         8  thrpt       10226.014          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue          128                        32  thrpt        9657.811          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue          128                       128  thrpt       10168.291          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue          128                       512  thrpt        9972.951          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue          512                         8  thrpt        7945.902          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue          512                        32  thrpt        7865.644          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue          512                       128  thrpt        7945.872          ops/ms
CustomBinaryDocValuesFieldBenchmark.binaryValue          512                       512  thrpt        7702.249          ops/ms

```